further adjustments with queue, enum, and message reply
