+++ config/__init__.py
@@ -1,8 +1,11 @@
 from __future__ import annotations
-__all__: list[str] = ["config_util", "data_class"]
+__all__: list[str] = ["config_util", "data_class", "default_cam_enum", "yam_read_error"]
 
 from . import config_util
 from . import data_class
+from . import default_cam_enum
+from . import yam_read_error
 from config.data_class import Open_Door_Task, Message_Task, Camera_Task
-
+from config.default_cam_enum import DefaultCam
+from config.yam_read_error import YamlReadError
