+++ config/data_class.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 from dataclasses import dataclass
-
+import telebot
 
 @dataclass
 class Open_Door_Task:
-    message_id: int = None
+    message: telebot.types.Message = None
     chat_id: int = None
     open: bool = False
     reply: bool =  False
@@ -14,7 +14,7 @@ class Open_Door_Task:
 @dataclass
 class Message_Task:
     chat_id: int = None
-    message_id: int = None
+    message: telebot.types.Message = None
     data_text: str = None
     filename: str = None
     send: bool = False
@@ -25,7 +25,7 @@ class Message_Task:
 @dataclass
 class Camera_Task:
     chat_id: int = None
-    message_id: int = None
+    message: telebot.types.Message = None
     reply: bool = False
     photo: bool = False
     blink_photo: bool = False
