+++ config/default_cam_enum.py
@@ -1,0 +1,5 @@
+from enum import Enum
+class DefaultCam(str, Enum):
+    """A custom enumeration that is YAML-safe."""
+    PICAM = "picam"
+    BLINK = "BLINK"
